Fix QueryConn in db/queries
Before this commit the QueryConn wouldn't work. Now it does.
You need to pass a database_uri to QueryConn like so:

    from lytics.settings import DevelopmentConfig
    query_conn = QueryConn(DevelopmentConfig.DATABASE_URI)
    query_conn.get_expenditures()

For testing you do something similar:

    from lytics.settings import TestingConfig
    query_conn = QueryConn(TestingConfig.DATABASE_URI)
    query_conn.get_expenditures()

This fullfills the idea of encapsulating the queries in one easy to use
class. Many of the methods of the class are not working so those fixes
will be in the next commits.